+++ client/android/src/main/kotlin/ClientAndroidPlugin.kt
@@ -3,13 +3,18 @@ package dev.inmo.wishlist.client
 import android.app.Application
 import android.content.Context
 import android.content.res.Resources
+import dev.inmo.micro_utils.coroutines.subscribeLoggingDropExceptions
 import dev.inmo.micro_utils.startup.plugin.StartPlugin
 import dev.inmo.navigation.core.NavigationChain
+import dev.inmo.navigation.core.extensions.changesInSubTreeFlow
+import dev.inmo.navigation.core.extensions.findNodeInSubTree
 import dev.inmo.navigation.core.repo.NavigationConfigsRepo
 import dev.inmo.wishlist.features.auth.client.ServerUrlStorage
 import dev.inmo.wishlist.features.common.client.models.ViewConfig
 import dev.inmo.wishlist.features.ui.scaffold.ui.ScaffoldViewConfig
 import dev.inmo.wishlist.features.ui.serverUrl.ui.ServerUrlViewConfig
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.merge
 import kotlinx.serialization.json.JsonObject
 import org.koin.core.Koin
 import org.koin.core.module.Module
@@ -40,13 +45,15 @@ class ClientAndroidPlugin(
         super.startPlugin(koin)
         val rootChain = koin.get<NavigationChain<ViewConfig>>()
         val storage = koin.get<ServerUrlStorage>()
-        val savedUrl = storage.getServerUrl()
-        if (savedUrl.isNullOrBlank()) {
-            if (rootChain.stackFlow.value.none { it.config is ServerUrlViewConfig }) {
-                rootChain.push(ServerUrlViewConfig())
+        val scope = koin.get<CoroutineScope>()
+        rootChain.changesInSubTreeFlow().subscribeLoggingDropExceptions(scope) {
+            val scaffoldView = rootChain.findNodeInSubTree { it.config is ScaffoldViewConfig }
+            val savedUrl = storage.getServerUrl()
+            if (scaffoldView != null && savedUrl == null) {
+                scaffoldView.chain.push(
+                    ServerUrlViewConfig()
+                )
             }
-        } else if (rootChain.stackFlow.value.none { it.config is ScaffoldViewConfig }) {
-            rootChain.push(ClientPlugin.mainScaffoldConfig)
         }
     }
 }